revert: deletion of L1 chain in N-API
