+++ crates/edr_napi/index.d.ts
@@ -197,6 +197,8 @@ export interface DebugTraceLogItem {
 }
 export const GENERIC_CHAIN_TYPE: string
 export function genericChainProviderFactory(): ProviderFactory
+export const L1_CHAIN_TYPE: string
+export function l1ProviderFactory(): ProviderFactory
 /** Identifier for the Ethereum spec. */
 export const enum SpecId {
   /** Frontier */
