+++ crates/edr_napi/src/spec.rs
@@ -1,4 +1,5 @@
 use edr_eth::{
+    chain_spec::L1ChainSpec,
     result::InvalidTransaction,
     transaction::{IsEip155, IsEip4844, TransactionMut, TransactionType, TransactionValidation},
 };
@@ -80,6 +81,39 @@ pub trait SyncNapiSpec:
     fn cast_response(
         response: Result<ResponseWithTraces<Self>, ProviderError<Self>>,
     ) -> napi::Result<Response>;
+}
+
+impl SyncNapiSpec for L1ChainSpec {
+    const CHAIN_TYPE: &'static str = "L1";
+
+    fn cast_response(
+        response: Result<ResponseWithTraces<Self>, ProviderError<Self>>,
+    ) -> napi::Result<Response> {
+        let response = jsonrpc::ResponseData::from(response.map(|response| response.result));
+
+        serde_json::to_string(&response)
+            .and_then(|json| {
+                // We experimentally determined that 500_000_000 was the maximum string length
+                // that can be returned without causing the error:
+                //
+                // > Failed to convert rust `String` into napi `string`
+                //
+                // To be safe, we're limiting string lengths to half of that.
+                const MAX_STRING_LENGTH: usize = 250_000_000;
+
+                if json.len() <= MAX_STRING_LENGTH {
+                    Ok(Either::A(json))
+                } else {
+                    serde_json::to_value(response).map(Either::B)
+                }
+            })
+            .map_err(|error| napi::Error::new(Status::GenericFailure, error.to_string()))
+            .map(|data| Response {
+                solidity_trace: None,
+                data,
+                traces: Vec::new(),
+            })
+    }
 }
 
 impl SyncNapiSpec for GenericChainSpec {
